[ch31343] [Intermittent] Continuous loading PD/SSFA Data in Programme Document
