+++ frontend_ip/src_ts/elements/ip-reporting/pd-details-reports.ts
@@ -8,13 +8,11 @@ import '../ip-reporting/pd-reports-toolbar';
 import '../ip-reporting/pd-reports-list';
 import {tableStyles} from '../../etools-prp-common/styles/table-styles';
 import {GenericObject} from '../../etools-prp-common/typings/globals.types';
-import Endpoints from '../../endpoints';
 import {Debouncer} from '@polymer/polymer/lib/utils/debounce';
 import {timeOut} from '@polymer/polymer/lib/utils/async';
 import {EtoolsPrpAjaxEl} from '../../etools-prp-common/elements/etools-prp-ajax';
 import {pdReportsFetch} from '../../redux/actions/pdReports';
 import {computePDReportsUrl, computePDReportsParams} from './js/pd-details-reports-functions';
-import {pdFetch} from '../../redux/actions/pd';
 
 /**
  * @polymer
@@ -33,8 +31,6 @@ class PdDetailsReport extends ReduxConnectedElement {
       <iron-location query="{{query}}"> </iron-location>
 
       <iron-query-params params-string="{{query}}" params-object="{{queryParams}}"> </iron-query-params>
-
-      <etools-prp-ajax id="programmeDocuments" url="[[programmeDocumentsUrl]]"> </etools-prp-ajax>
 
       <etools-prp-ajax id="pdReports" url="[[pdReportsUrl]]" params="[[pdReportsParams]]"> </etools-prp-ajax>
 
@@ -64,21 +60,10 @@ class PdDetailsReport extends ReduxConnectedElement {
   @property({type: Object, computed: '_computePDReportsParams(pdId, queryParams)'})
   pdReportsParams!: GenericObject;
 
-  @property({type: String, computed: '_computeProgrammeDocumentsUrl(locationId)'})
-  programmeDocumentsUrl!: string;
-
-  @property({
-    type: Object,
-    computed: 'getReduxStateValue(rootState.programmeDocumentReports.countByPD)',
-    observer: '_getPdReports'
-  })
-  pdReportsCount!: GenericObject;
-
   @property({type: String, computed: 'getReduxStateValue(rootState.programmeDocumentReports.current.id)'})
   pdReportsId!: string;
 
   private _debouncer!: Debouncer;
-  private fetchPdsDebouncer!: Debouncer;
 
   public static get observers() {
     return ['_handleInputChange(pdReportsUrl, pdReportsParams)'];
@@ -90,10 +75,6 @@ class PdDetailsReport extends ReduxConnectedElement {
 
   _computePDReportsParams(pdId: string, queryParams: GenericObject) {
     return computePDReportsParams(pdId, queryParams);
-  }
-
-  _computeProgrammeDocumentsUrl(locationId: string) {
-    return locationId ? Endpoints.programmeDocuments(locationId) : '';
   }
 
   _handleInputChange(url: string) {
@@ -114,31 +95,6 @@ class PdDetailsReport extends ReduxConnectedElement {
           console.log(err);
         });
     });
-  }
-
-  _getPdReports() {
-    // Status being present prevents res.data from getting reports,
-    // preventing pd-details title from rendering. Deleting the status
-    // can resolve this issue, and filter will still work
-    if (this.pdReportsCount[this.pdId] > 0 && this.pdReportsId === '') {
-      this.fetchPdsDebouncer = Debouncer.debounce(this.fetchPdsDebouncer, timeOut.after(100), () => {
-        const pdThunk = this.$.programmeDocuments as EtoolsPrpAjaxEl;
-        pdThunk.params = {
-          page: 1,
-          page_size: 10,
-          programme_document: this.pdId
-        };
-
-        // Cancel the pending request, if any
-        (this.$.programmeDocuments as EtoolsPrpAjaxEl).abort();
-        this.reduxStore
-          .dispatch(pdFetch(pdThunk.thunk()))
-          // @ts-ignore
-          .catch(function (err) {
-            console.log(err);
-          });
-      });
-    }
   }
 }
